Damage effect fix :wrench:

+++ src/Projectiles/BasicProjectile.h
@@ -6,7 +6,7 @@
 class BasicProjectile : public Projectile{
 
 public:
-    BasicProjectile(vector3df pos, vector3df dir, int emi, int damageMult = 1);
+    BasicProjectile(vector3df pos, vector3df dir, int emi, float damageMult = 1);
 
 private:
     void ContactAction(Player* p);
